Fix GPU gear && misc..

+++ system/engine/gears/006pqr.sh
@@ -4,6 +4,7 @@
 B=/system/engine/bin/busybox
 LOG=/sdcard/Android/FDE.txt
 TIME=$($B date | $B awk '{ print $4 }')
+SDK=$(getprop ro.build.version.sdk)
 
 $B echo "" >> $LOG
 $B echo "[$TIME] 006 - ***GPU gear***" >> $LOG
@@ -18,15 +19,15 @@ if [ -e /system/lib/egl/libGLESv2_adreno200.so ]; then
   $B chmod 777 /system/engine/assets/adreno_config.txt
   $B ln -s /system/engine/assets/adreno_config.txt /data/local/tmp/adreno_config.txt
  fi;
- if [ -e /sys/devices/platform/kgsl/msm_kgsl/kgsl-3d0/io_fraction ]; then
-  $B echo 50 > /sys/devices/platform/kgsl/msm_kgsl/kgsl-3d0/io_fraction
- fi;
- if [ "$AV" -ne "1712" ]; then
-  $B echo "Forcing GPU to render UI.." >> $LOG
-  $B rm /system/lib/egl/libGLES_android.so
-  $B sed -i '/0 0 android/d' /system/lib/egl/egl.cfg
- else
-  $B echo "You have legacy adreno libs. No HWA for you." >> $LOG
+ if [ -e /system/lib/egl/libGLES_android.so ]; then
+  if [ "$AV" -ne "1712" ]; then
+   $B echo "Forcing GPU to render UI.." >> $LOG
+   $B mount -o remount,rw /system
+   $B rm /system/lib/egl/libGLES_android.so
+   $B sed -i '/0 0 android/d' /system/lib/egl/egl.cfg
+  else
+   $B echo "You have legacy adreno libs. No HWA for you." >> $LOG
+  fi;
  fi;
  $B mount -t debugfs debugfs /sys/kernel/debug
  $B echo "Setting correct device permissions.." >> $LOG
@@ -45,21 +46,32 @@ $B mount -t debugfs debugfs /sys/kernel/debug
 $B chmod 777 /dev/graphics/fb0
 if [ -e /sys/kernel/debug/msm_fb/0/vsync_enable ]; then
  $B echo "Disabling VSYNC.." >> $LOG
- $B chmod 644 /sys/kernel/debug/msm_fb/0/vsync_enable
  $B echo 0 > /sys/kernel/debug/msm_fb/0/vsync_enable
 fi;
-
-if [ -e /system/xbin/sqlite3 ]; then
- $B echo "Tuning Android animations.." >> $LOG
- echo "REPLACE INTO \"system\" VALUES(26,'window_animation_scale','0.25');REPLACE INTO \"system\" VALUES(27,'transition_animation_scale','0.25');" | sqlite3 /data/data/com.android.providers.settings/databases/settings.db
- $B kill `$B pidof system_server`
+if [ -e /sys/devices/platform/kgsl/msm_kgsl/kgsl-3d0/io_fraction ]; then
+ $B echo 50 > /sys/devices/platform/kgsl/msm_kgsl/kgsl-3d0/io_fraction
+fi;
+if [ "$SDK" -le "17" ]; then
+ if [ -e /system/engine/prop/firstboot ]; then
+  if [ -e /system/xbin/sqlite3 ]; then
+   $B echo "Tuning Android animations.." >> $LOG
+   $B echo "REPLACE INTO \"system\" VALUES(26,'window_animation_scale','0.25');REPLACE INTO \"system\" VALUES(27,'transition_animation_scale','0.25');" | sqlite3 /data/data/com.android.providers.settings/databases/settings.db
+  fi;
+ fi;
+fi;
+if [ -e /sys/module/tpd_setting/parameters/tpd_mode ]; then
+ $B chmod 644 /sys/module/tpd_setting/parameters/tpd_mode
+ $B echo 1 > /sys/module/tpd_setting/parameters/tpd_mode
+elif [ -e /sys/module/hid_magicmouse/parameters/scroll_speed ]; then
+ $B chmod 644 /sys/module/hid_magicmouse/parameters/scroll_speed
+ $B echo 63 > /sys/module/hid_magicmouse/parameters/scroll_speed
 fi;
 
 $B echo "Tuning Android graphics.." >> $LOG
 setprop debug.sf.hw 1
 setprop debug.egl.hw 1
 setprop debug.egl.profiler 1
-setprop debug.egl.swapinterval -60
+setprop debug.egl.swapinterval 1
 setprop debug.gr.numframebuffers 3
 setprop debug.gr.swapinterval 1
 setprop debug.mdpcomp.logs 0
@@ -76,9 +88,6 @@ setprop video.accelerate.hw 1
 setprop windowsmgr.max_events_per_sec 90
 setprop windowsmgr.support_rotation_270 true
 setprop hwui.render_dirty_regions false
-setprop persist.sys.NV_FPSLIMIT 90
-setprop persist.sys.NV_POWERMODE 1
-setprop persist.sys.NV_POWER_MODE 1
 setprop persist.sys.scrollingcache 3
 setprop ro.media.dec.jpeg.memcap 8000000
 setprop ro.media.enc.hprof.vid.bps 8000000
@@ -96,14 +105,7 @@ setprop persist.service.pcsync.enable false
 setprop ro.lge.proximity.delay 25
 setprop mot.proximity.delay 25
 setprop ro.telephony.call_ring.delay 0
-
-if [ -e /sys/module/tpd_setting/parameters/tpd_mode ]; then
-$B chmod 644 /sys/module/tpd_setting/parameters/tpd_mode
-$B echo 1 > /sys/module/tpd_setting/parameters/tpd_mode
-elif [ -e /sys/module/hid_magicmouse/parameters/scroll_speed ]; then
-$B chmod 644 /sys/module/hid_magicmouse/parameters/scroll_speed
-$B echo 63 > /sys/module/hid_magicmouse/parameters/scroll_speed
-fi;
+persist.sys.strictmode.disable true
 
 TIME=$($B date | $B awk '{ print $4 }')
 $B echo "[$TIME] 006 - ***GPU gear*** - OK" >> $LOG
